+++ __tests__/components/header.test.jsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Header from '../components/header';
+import Header from '../../components/header';
 
 describe('Header', () => {
     it('Renders the site title', () => {
